Move function to a module
